Add filter searching on in need and categories
We need to be able to search on in need to find businesses that are in
need and on categories to find well... different businesses?

Anyway its all in there. THe URL structures are pre defined, though all
on its own seems to be failing so I will take a look at that.