Add search load state, fix no results flash and images
Basically a tidying things up that were annoying me commit.

Added a loading state to the spinner, seemed to come in the UI library
so was pretty easy to add.

Fixed the no results flash so that it will only show we we have
confirmed that there really are no results.

Resized the images to an appropriate size and set DPR to 2 so they can
be retina.